requirements.txt
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyPDF2==2.11.0
python-dotenv==0.21.0
